pyproject.toml
[tool.poetry]
name = "kpi-new"
version = "0.1.0"
description = ""
authors = ["restevean <58455099+restevean@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
boto3 = "^1.35.49"
pysftp = "^0.2.9"
pandas = "^2.2.3"
tabulate = "^0.9.0"
pytest-mock = "^3.14.0"
jsonmerge = "^1.9.2"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
